fix: report 1-based line numbers consistently in syntax errors
The lexer counted lines from 0 and only one parser error compensated
for it, so different errors reported different numbers for the same
line. Also fixes the swapped include guards of token_.hpp/lexer.hpp,
drops unused includes, and uses size_t for the lexer cursor to silence
sign-compare warnings.

+++ src/lexer.cpp
@@ -1,6 +1,5 @@
 #include "../include/lexer.hpp"
-#include <iostream>
-#include <fstream>
+#include <cctype>
 
 char Lexer::peek() {
     if (curr + 1 >= file.size()) return '\0';
@@ -15,7 +14,7 @@ void Lexer::consume() {
     curr++;
 }
 
-int Lexer::index() {
+size_t Lexer::index() {
     return curr;
 }
